% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/OCC.R
\name{hyp_occurve}
\alias{hyp_occurve}
\title{OCC Function to create OC Curves - Harish Jose}
\usage{
hyp_occurve(c = 0, n = 60, N = 1200, alpha = 5, beta = 10)
}
\description{
This function allows you to create OC Curves based on Hypergeometric Distribution.
You need to enter the folowing details.
c = number of defects (default is 0)
n = sample size (default is 60)
N = Lot Size (defaut is 1200)
alpha = level of significance or Producer's risk (default is 5%)
beta = consumer's risk (default is 10%)
Enter alpha and beta as values between 1 and 100
#'
}
\examples{
hyp_occurve()
}
\keyword{Curve}
\keyword{OC}

