\name{mergeCTD}
\alias{mergeCTD}
\title{
Combine CTD casts
}
\description{
Concatenate individual CTD casts from a region or island into a single Excel file for the database.
}
\usage{
mergeCTD(dir, cruise, group)
}

\arguments{
  \item{dir}{Folder containing CTD casts
}
  \item{cruise}{Name of cruise
}
  \item{group}{Grouping variable, either region (e.g. "PRIA") or island (e.g. "JAR")
}
}
