\name{processCTD_profile}
\alias{processCTD_profile}

\title{Process CTD profiles
}
\description{Batch processes SBE19, SBE19plus, and SBE19plusV2 CNV files in a directory to CSV.}
\usage{
processCTD_profile(dir)
}
\arguments{
  \item{dir}{Folder path to CNV files
}
}
\details{CTD profiles must be converted from HEX to CNV in SeaTerm software prior to processing.
}

