\name{processSTR}
\alias{processSTR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Clean STR time series
}
\description{Launches a GUI that allows user to select start and end times of STR time series and exports to CDP file.
}
\usage{
processSTR(file, write.csv = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{STR CSV file
}
  \item{write.csv}{Specify \code{TRUE} to save time series as CSV (in addition to CDP)
}


}
