\name{processSTR}
\alias{processSTR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Clean STR time series
}
\description{Launches a GUI that allows user to select start and end times of STR time series and exports to CDP file. Can either process a SBE logger data file (exported to CSV) or RBR logger data file (downloaded as RSK).
}
\usage{
processSTR(file, model = c("SBE","RBR"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{STR CSV file
}
  \item{model}{Specify "SBE" or "RBR"
}



}
